Change AreaInHectares from double to decimal for financial precision
Decimal is the correct type for area calculations involving monetary values,
avoiding floating-point rounding errors in price computations.

+++ AgriMarket.BLL/Dtos/Bookings/BookingDto.cs
@@ -8,7 +8,7 @@ public sealed class BookingDto
     public Guid Id { get; init; }
     public BookingStatus Status { get; init; }
     public decimal TotalPrice { get; init; }
-    public double AreaInHectares { get; init; }
+    public decimal AreaInHectares { get; init; }
     public DateTime CreatedAt { get; init; }
     public string? Notes { get; init; }
     public Guid ServiceListingId { get; init; }
