+++ AgriMarket.BLL/Dtos/Bookings/BookingSummaryDto.cs
@@ -7,7 +7,7 @@ public sealed class BookingSummaryDto
     public Guid Id { get; init; }
     public string ClientName { get; init; } = default!;
     public BookingStatus Status { get; init; }
-    public double AreaInHectares { get; init; }
+    public decimal AreaInHectares { get; init; }
     public decimal TotalPrice { get; init; }
     public DateTime CreatedAt { get; init; }
 }
