+++ AgriMarket.Web/Areas/Client/ViewModels/Bookings/BookingDetailsViewModel.cs
@@ -7,7 +7,7 @@ public class BookingDetailsViewModel
     public Guid Id { get; set; }
     public BookingStatus Status { get; set; }
     public decimal TotalPrice { get; set; }
-    public double AreaInHectares { get; set; }
+    public decimal AreaInHectares { get; set; }
     public DateTime CreatedAt { get; set; }
     public string? Notes { get; set; }
     public string ListingTitle { get; set; } = default!;
