+++ AgriMarket.Web/Areas/Client/ViewModels/Bookings/CreateBookingViewModel.cs
@@ -12,7 +12,7 @@ public class CreateBookingViewModel
 
     [Required]
     [Range(0.1, 10000, ErrorMessage = "Area must be between 0.1 and 10000 hectares")]
-    public double AreaInHectares { get; set; }
+    public decimal AreaInHectares { get; set; }
 
     [MaxLength(1000)]
     public string? Notes { get; set; }
