+++ AgriMarket.Web/Areas/Client/ViewModels/MyListings/BookingsForListingItemViewModel.cs
@@ -7,7 +7,7 @@ namespace AgriMarket.Web.Areas.Client.ViewModels.MyListings
         public Guid Id { get; set; }
         public string ClientName { get; set; } = default!;
         public BookingStatus Status { get; set; }
-        public double AreaInHectares { get; set; }
+        public decimal AreaInHectares { get; set; }
         public decimal TotalPrice { get; set; }
         public DateTime CreatedAt { get; set; }
     }
